refactor: コンポーネント構成ルールの統一とファイル構造の最適化
- ファイル命名規則を統一（*.container.tsx → container.tsx, *.presentational.tsx → presentational.tsx）
- Container/Presentational パターンの役割分離を明確化
  - _containers/container.tsx: Server Componentでデータ取得
  - _containers/presentational.tsx: Client ComponentでUI・ユーザー操作
- 不適切な配置を修正（_components/presentational.tsx → _containers/presentational.tsx）
- 全ルーティング機能セグメントに統一構造を適用
- テストファイルの配置とインポートパスを最適化
- CLAUDE.mdにコンポーネント構成ルールを文書化

品質チェック:
- npm run lint: エラーなし
- npm run test: 87テスト全てパス
- npm run build: 成功
- Playwright動作確認: 全機能正常動作
