+++ apps/frontend/app/features/routing/basic/_containers/container.tsx
@@ -1,0 +1,11 @@
+import { BasicRoutingPresentational } from './presentational';
+
+export function BasicRoutingContainer() {
+  // Server Componentでデータ取得
+  const serverData = {
+    timestamp: new Date().toISOString(),
+    serverRenderTime: Date.now()
+  };
+
+  return <BasicRoutingPresentational serverData={serverData} />;
+}
